requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tabulate==0.9.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask==3.0.3
flask-cors==4.0.0
